refactor(veille): drop dashboard/deliveries écrans + extract VeilleSourceExample (Story 23.2, commit 1/5)
Commit 1 de PR-4 (refonte UX veille intégrée) — drop des écrans morts qui
résultaient de la suppression backend du pipeline LLM async (Story 23.1).
Ces écrans appelaient des endpoints désormais 410 Gone.

Drops :
- veille_dashboard_screen.dart (659 L)
- veille_deliveries_screen.dart (226 L)
- veille_delivery_detail_screen.dart (239 L)
- widgets/veille_cluster_card.dart (126 L)
- providers/veille_deliveries_provider.dart (49 L)
- test/screens/veille_dashboard_screen_test.dart

Routes & deep links :
- Drop routes veilleDashboard, veilleDeliveries, veilleDeliveryDetail
- deep_link_service : redirect veille → /flux-continu (où vit désormais
  le contenu veille comme slot de la Tournée du jour)
- push_notification_service : payload notif veille → /flux-continu
- settings_sheet : drop l'entrée "Configurer ma veille" (entry point
  désormais via Mes intérêts, commits 4-5)

Extraction :
- VeilleSourceExample extraite de veille_delivery.dart vers
  models/veille_source_example.dart (gardée pour step3, alors que
  veille_delivery.dart sera dropé en commit 2).

veille_config_screen.dart : minimal patch — redirects post-submit
vers /flux-continu (au lieu de veilleDashboard/Deliveries supprimés).
Refonte complète du submit-async en commit 2.

Verify :
- flutter analyze : 0 erreur (warnings pré-existants non liés)
- flutter test test/core/services/deep_link_service_test.dart
  test/features/veille/widgets/veille_source_card_test.dart → 16/16 ✓

Story doc : docs/stories/core/23.2.veille-mobile-favori-tournee-integration.md
