+++ docs/stories/core/23.2.veille-mobile-favori-tournee-integration.md
@@ -1,0 +1,78 @@
+# 23.2 — Veille mobile : intégration favori + Tournée du jour (PR-4)
+
+> Refonte UX mobile de la veille, **révisant le PR-3 original de la Story 23.1**. La veille n'est plus un espace produit séparé : elle devient le 3ᵉ type de favori dans "Mes intérêts" et un slot stylisé de la Tournée du jour. Plus aucun écran dédié (`/veille/dashboard`, `/veille/deliveries`).
+
+**Status** : 🚧 En cours — plan PO validé 2026-05-19, exécution démarrée 2026-05-19.
+**Type** : Feature + refactor (réduction de complexité côté UX).
+**Plan source** : `/Users/laurinboujon/.claude/plans/system-instruction-you-are-working-sprightly-bengio.md` (file approuvé via ExitPlanMode 2026-05-19).
+**Dépend de** : Story 23.1 PR-1/2 mergés (scheduler stoppé, feed filtré live, endpoints LLM 410), PR-3 (extension `UserFavoriteInterest.veille_config_id` mergé en #632).
+
+---
+
+## Vision révisée (PO, 2026-05-19)
+
+L'original PR-3 de Story 23.1 prévoyait un écran feed veille dédié (`/veille`). La vision a pivoté :
+
+- **Entry point** : la page "Mes intérêts" (CTA "Crée ta veille" + favori veille avec badge), **plus de "Configurer ma veille"** dans le settings sheet.
+- **Affichage du contenu** : un slot `kind=veille` dans la Tournée du jour (`flux-continu`), avec accent visuel `sectionVeille1/2`.
+- **Flow config** : 5 écrans (intro → step1 → step1.5 preset → step2 → step3), step4 frequency drop, **skip** sur step2 + **mode "Configuration avancée"** sur step2/step3.
+- **Suppression** : dashboard, deliveries, delivery detail, transitions, step4.
+
+---
+
+## Décisions tranchées (audit + PO)
+
+| Sujet | Décision | Notes |
+|---|---|---|
+| Force-update mobile | **Gestion gracieuse 410 Gone** via intercepteur Dio global → banner "Mise à jour requise" | Pas de mécanisme `min_app_version` côté backend ; construire ça est hors scope. |
+| Deep link `veille/dashboard` | Redirige vers Tournée (`/flux-continu`) | C'est là où vit le contenu veille désormais. |
+| Intro screen | Conditionnelle sur 404 de `getConfig()` | Pas de flag SharedPreferences ; première fois = pas de config. |
+| `VeilleSourceExample` | Extraite vers `models/veille_source_example.dart` | Couplage caché : utilisée par step3 via `veille_source_examples_provider`. |
+| Découpage | 5 commits intermédiaires dans 1 seule PR | Atomicité produit, chaque commit laisse `flutter analyze` clean. |
+
+---
+
+## Plan d'exécution (5 commits)
+
+| # | Périmètre | Statut |
+|---|---|---|
+| 1 | Drop dashboard/deliveries/delivery_detail + cluster_card + deliveries_provider. Extract `VeilleSourceExample`. Clean routes + deep_links + settings sheet. Minimal patch `veille_config_screen` (redirects → fluxContinu). | 🚧 En cours |
+| 2 | Refonte provider (931→~300 L) + repo (drop suggestions/deliveries, add getFeed) + DTO (drop frequency/dayOfWeek, add keywords) + widgets (1502→~800 L). Drop step4/flow_loading/poll/halo_loader/delivery/suggestion/suggestions_provider/mock_data. Refonte `veille_config_screen` (drop submit-async). | ⏳ Pending |
+| 3 | Refonte step2 (skip + advanced) + step3 (advanced URL source) + intro conditionnelle. | ⏳ Pending |
+| 4 | `VeilleFavoriteRef` sealed class + CTA "Crée ta veille" dans Mes intérêts + menu Modifier/Archiver. | ⏳ Pending |
+| 5 | `SectionKind.veille` + slot Tournée stylisé + intercepteur Dio 410 Gone. | ⏳ Pending |
+
+---
+
+## Fichiers critiques (chemins absolus)
+
+```
+apps/mobile/lib/features/veille/providers/veille_config_provider.dart
+apps/mobile/lib/features/veille/repositories/veille_repository.dart
+apps/mobile/lib/features/veille/models/veille_config_dto.dart
+apps/mobile/lib/features/veille/screens/veille_config_screen.dart
+apps/mobile/lib/features/veille/screens/steps/step2_suggestions_screen.dart
+apps/mobile/lib/features/veille/screens/steps/step3_sources_screen.dart
+apps/mobile/lib/features/veille/widgets/veille_widgets.dart
+apps/mobile/lib/features/my_interests/models/user_interests_state.dart
+apps/mobile/lib/features/my_interests/screens/my_interests_screen.dart
+apps/mobile/lib/features/my_interests/widgets/favorites_reorderable_section.dart
+apps/mobile/lib/features/flux_continu/models/flux_continu_models.dart
+apps/mobile/lib/features/flux_continu/providers/flux_continu_provider.dart
+apps/mobile/lib/features/flux_continu/screens/flux_continu_screen.dart
+apps/mobile/lib/config/routes.dart
+apps/mobile/lib/core/services/deep_link_service.dart
+apps/mobile/lib/features/settings/widgets/settings_sheet.dart
+```
+
+---
+
+## Vérification
+
+- `cd apps/mobile && flutter test && flutter analyze` 100% vert et 0 warning.
+- `/validate-feature` exécuté avec `.context/qa-handoff.md` rédigé.
+- 3 scénarios Playwright : happy path création (avec skip step2), mode advanced (keywords + URL niche), édition + archive.
+
+## Change Log
+
+- 2026-05-19 — Story créée, commit 1 démarré.
